Async consuming of messages to avoid delayed action on candles.
The big assumption is that candles can be processed independently. Why
then preserve the order by using a queue? The state machine needs the
ordered processing. So the only way out is to reduce the 1second
bottleneck from the greeks server or process candles at intervals like 2
or 5 seconds, providing enough time for the processing to complete.